+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
  * Deprecate the `reset_on_message` config option. It can be set to `true` only and does nothing now
  * Add `trust_x_sendfile_type_header` option
  * Add support for first-class callable route controller in `MicroKernelTrait`
+ * Add tag `routing.condition_service` to autoconfigure routing condition services.
 
 6.0
 ---